server/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.27.2
websockets==12.0
faster-whisper
soundfile
numpy
piper-tts
